CRUD Added for Cart and CartItem Controller

+++ src/main/java/com/entity/CartEntity.java
@@ -29,6 +29,6 @@ public class CartEntity {
   @JoinColumn(name = "restaurantId")
   private RestaurantEntity restaurant;
 
-  Integer offerAmount;
+  Double offerAmount;
 
 }
